requirements/requirements.txt
python-telegram-bot[job-queue]
PlexAPI
requests
backoff
pyinstaller
pyautogui
pycaw